# specs/001-email-scan-notify/contracts/api.yaml
openapi: 3.1.0
info:
  title: Email Scan & Notify API
  description: |
    REST API for the Email Scan & Notify service.
    Connects to Gmail via OAuth2, scans emails in the background,
    and delivers notifications to Slack and/or WhatsApp channels.
  version: 1.0.0
  contact:
    name: Engineering Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/notification-channels:
    post:
      operationId: addNotificationChannel
      tags: [Notification Channels]
      summary: Add a notification channel (Slack or WhatsApp)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationChannelRequest"
      responses:
        "201":
          description: Channel created
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listNotificationChannels
      tags: [Notification Channels]
      summary: List configured notification channels
      responses:
        "200":
          description: List of channels
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationChannel"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/notification-channels/{id}:
    patch:
      operationId: updateNotificationChannel
      tags: [Notification Channels]
      summary: Update a notification channel configuration
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationChannelRequest"
      responses:
        "200":
          description: Channel updated
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteNotificationChannel
      tags: [Notification Channels]
      summary: Remove a notification channel
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Channel deleted
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---- Filter Rules -------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---- Reusable Parameters ------------------------------------------------
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique resource identifier

  # ---- Reusable Headers ---------------------------------------------------
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
    X-RateLimit-Reset:
      description: Unix epoch timestamp when the rate-limit window resets
      schema:
        type: integer

  # ---- Reusable Responses -------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---- Schemas ------------------------------------------------------------
  schemas:
...
    ChannelType:
      type: string
      enum: [slack, whatsapp]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              examples: ["VALIDATION_ERROR"]
            message:
              type: string
              description: Human-readable error description
              examples: ["Email address is required"]
            details:
              type: object
              additionalProperties: true
              description: Optional structured details

    # -- Pagination ---------------------------------------------------------
...
    CreateNotificationChannelRequest:
      type: object
      required: [type, config]
      properties:
        type:
          $ref: "#/components/schemas/ChannelType"
        label:
          type: string
          maxLength: 128
          description: Human-friendly label for the channel
        config:
          $ref: "#/components/schemas/ChannelConfig"
    UpdateNotificationChannelRequest:
      type: object
      properties:
        label:
          type: string
          maxLength: 128
        config:
          $ref: "#/components/schemas/ChannelConfig"
        enabled:
          type: boolean
    ChannelConfig:
      type: object
      description: Configuration varies by channel type
      properties:
        slack_webhook_url:
          type: string
          format: uri
          description: Required when type is "slack"
        slack_channel:
          type: string
          description: Optional Slack channel override
        whatsapp_phone_number:
          type: string
          description: Required when type is "whatsapp" (E.164 format)
          examples: ["+14155551234"]
    NotificationChannel:
      type: object
      required: [id, type, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/ChannelType"
        label:
          type: string
        config:
          $ref: "#/components/schemas/ChannelConfig"
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # -- Filter Rules -------------------------------------------------------
...